Swift_inventory script now concats policy data
The swift_inventory script now combines the required variables to
generate the storage policy definitions into a concatenated string
and stored with the host.

  extra_data=<name>:<index>:<depricated>:<type>

E.G:

  extra_data=gold:0:True:Replication

If the host is a member of more then one storage group then these
entries are concatenated together with a ';':

  extra_data=<name1>:<index1>:..:<type>;<name2>:<index2>....

The swift.conf file template now takes this into account and
is generated correctly.