Make ring_builder understand the conf_d parameter
This change allows the ring_builder to follow the conf_d parameter
and load in conf.d files. This will allow us to point to the main
rpc_user_config.yml and build swift rings no matter where a user/op
may add swift definitions.

It loads in the conf_d parameter, and if it exists, it will search for
the folder:

  - If it looks like a full path (starts with '/') it will attempt to
    use it.

  - If it looks like a relative path then:
        * it will first look for it  being relative to the current
          working directory.

        * then look for the conf_d directory relative to the config file
          passed in.